the frontend has successfully made CRUD commands throught the PHP API.
Next the app will need to update the page async. I need to figure out where the partials for these page changes will live. But before that I will be refactor the code base a bit. It seems that it may be time to split the gsd project off into its own repo. For now I will move the dependent files and dirs into the build/gsd/ as that is the root of this project. I think there might be a potential to use .gitignore to put that dir into its own repo.